+++ CHANGELOG.md
@@ -23,3 +23,5 @@
 * Define `Adj.Program.Construction` module
 * Define `:.` method in `Categroy` typeclass
 * Define `Adj.Program.Option` module
+* Define `Covariant` type family
+* Define `Contravariant` type family